+++ src/main/java/com/github/osisoft/ocs_sample_library_preview/dataviews/Field.java
@@ -1,14 +1,32 @@
 package com.github.osisoft.ocs_sample_library_preview.dataviews;
 
+import com.github.osisoft.ocs_sample_library_preview.*;
+import com.github.osisoft.ocs_sample_library_preview.sds.*;
+
 public class Field {
     private FieldSource Source;
     private String[] Keys;
     private String Label;
+    private boolean IncludeUom;
+    private SummaryDirection SummaryDirection;
+    private SdsSummaryType SummaryType;
 
     public Field(FieldSource source, String[] keys, String label) {
         this.Source = source;
         this.Keys = keys;
         this.Label = label;
+        this.IncludeUom = false;
+        this.SummaryDirection = SummaryDirection.Forward;
+        this.SummaryType = SdsSummaryType.None;
+    }
+
+    public Field(Field field) {
+        this.Source = field.getSource();
+        this.Keys = field.getKeys();
+        this.Label = field.getLabel();
+        this.IncludeUom = field.getIncludeUom();
+        this.SummaryDirection = field.getSummaryDirection();
+        this.SummaryType = field.getSummaryType();
     }
 
     public FieldSource getSource() {
@@ -34,4 +52,18 @@ public class Field {
     public void setLabel(String label) {
         this.Label = label;
     }
+
+    public boolean getIncludeUom() {
+        return IncludeUom;
+    }
+
+    public void setIncludeUom(boolean includeUom) { this.IncludeUom = includeUom; }
+
+    public SummaryDirection getSummaryDirection() { return SummaryDirection; }
+
+    public void setSummaryDirection(SummaryDirection summaryDirection) { this.SummaryDirection = summaryDirection; }
+
+    public SdsSummaryType getSummaryType() { return SummaryType; }
+
+    public void setSummaryType(SdsSummaryType summaryType) { this.SummaryType = summaryType; }
 }