+++ HISTORY.md
@@ -1,5 +1,12 @@
 # Version History
 
+## 0.0.27-preview / 2021-05-24
+
+- Added IncludeUom property to Field class
+- Added SummaryDirection to Field class and enum to package
+- Added SdsSummaryType to Field class and enum to package
+- Added Uom to Sds Property class
+
 ## 0.0.26-preview / 2021-05-19
 
 - Updated dependencies